* fixed several memory leaks

+++ src/ta3d/src/3do.cpp
@@ -170,7 +170,7 @@ namespace TA3D
 
 	void ANIMATION_DATA::destroy()
 	{
-		for (byte i = 0; i < 3; ++i)
+        for (int i = 0; i < 3; ++i)
 		{
 			if (axe[i])
 				delete[] axe[i];
